Add test script
