+++ src/raspi/component/motor/__init__.py
@@ -1,2 +1,3 @@
 from . import raw_motor as raw_motor_action
+from . import odrive_motor as odrive_motor_action
 from .pin import *
