+++ src/raspi/pinconfig.json
@@ -107,5 +107,17 @@
 			"left_limit_switch": "io_expander_1.channel_4",
 			"right_limit_switch": "io_expander_1.channel_5"
 		}
+	},
+	"can_bus": {
+		"can_bus_1": {
+			"channel": "can0",
+			"interface": "socketcan"
+		}
+	},
+	"odrive_motor": {
+		"odrive_1": {
+			"bus": "can_bus_1",
+			"axisID": 0
+		}
 	}
 }
