+++ src/raspi/samples/can/cantools_testing.py
@@ -1,4 +1,5 @@
 import cantools
+from enum import IntEnum
 
 db = cantools.database.load_file("src/raspi/odrive-cansimple.dbc")
 
@@ -8,7 +9,16 @@ heartbeat_data = {'Trajectory_Done_Flag': 0,
                     'Encoder_Error_Flag': 0, 
                     'Motor_Error_Flag': 0,
                     'Axis_State': 3,
-                    'Axis_Error': 12}
+                    'Axis_Error': 128}
 encoded_data = heartbeat_msg.encode(heartbeat_data)
 print(encoded_data)
-print(type(heartbeat_msg.decode(encoded_data)['Axis_State']))
+print(heartbeat_msg.decode(encoded_data)['Axis_Error'] is not int)
+
+class Test(IntEnum):
+    woah = 2
+
+def Hello(a: Test):
+    print(a)
+
+Hello(2)
+Hello(Test.woah)
